pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "dreams-rns-rocm-lumi"
version = "0.1.0"
description = "GPU-accelerated Ramanujan Dreams pipeline using RNS arithmetic on AMD MI250X via ROCm/HIP"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "David Svensson"},
]

dependencies = [
    "numpy>=1.24.0",
    "sympy>=1.12",
    "pyyaml>=6.0",
    "mpmath>=1.3.0",
]

[project.optional-dependencies]
mpi = ["mpi4py>=3.1.0"]
dev = ["pytest>=7.0", "pytest-timeout"]

[tool.setuptools.packages.find]
include = ["dreams_rocm*"]

[tool.pytest.ini_options]
testpaths = ["dreams_rocm/rns/tests"]
python_files = "test_*.py"
